+++ packages/comet-admin-rte/CHANGELOG.md
@@ -8,10 +8,10 @@ All notable changes to this project will be documented in this file. This projec
 
 -   Add default styles (MUI) to built-in blocktypes
 -   Make built-in blocktypes styleable
+-   Rename prop-name "customBlockMap" to "blocktypeMap"
+-   Deprecate prop-name "customBlockMap"
 
 ### Incompatible Changes
-
--   Rename prop-name "customBlockMap" to "blocktypeMap"
 
 ### Internal Changes